feat(rust): add extract_ordered_extractions record walker
Fifth increment of the Rust port. Ports the second half of
Resolver.resolve() — specifically extract_ordered_extractions — from
langextract/resolver.py. This is the bridge between
FormatHandler::parse_output (which produces a sequence of flat
key-value records) and the higher-level pipeline (which wants
Vec<Extraction>).

The implementation lives in langextract-format::records, not in a
new langextract-resolver crate. The conversion is pure "further
processing of parsed records" with no tokenization or alignment
concerns, so it naturally belongs with parse_output as its
companion. A consumer can now do:

    let records = handler.parse_output(llm_output, None)?;
    let extractions = extract_ordered_extractions(&records, "_attributes", Some("_index"))?;
    let grounded = align_extraction_groups(vec![extractions], source, &opts)?;

and get fully-grounded Extractions from raw LLM text end to end.

Two supporting changes to langextract-core:

- Extraction::extraction_index widens from Option<usize> to
  Option<i64>. Python's extraction_index is an unbounded int and the
  existing test suite exercises negative values ("lower priority" or
  "placeholder" records). Signed i64 is the smallest type that
  preserves Python semantics and round-trips cleanly through
  serde_json's Number::as_i64.

- langextract-format depends on serde_json with the preserve_order
  feature. This is load-bearing for duplicate-index ordering: when
  two extractions in the same record share an extraction_index, the
  stable sort keeps them in source order, which only works if the
  parsed ParsedRecord iterates in insertion order. Without
  preserve_order, serde_json's default alphabetic Map iteration
  would silently reorder the output.

Records module details:

- extract_ordered_extractions walks each record in order, emitting
  one Extraction per non-suffix field that resolves to a scalar.
- Keys ending in the index suffix are validated as integers and
  skipped (they're metadata).
- Keys ending in the attribute suffix are validated as objects or
  null and skipped (they're attached via lookup).
- Two ordering modes: explicit index_suffix ("_index") reads the
  paired index field and drops extractions whose class has no index
  key (matches Python's silent-drop behaviour), or None to
  auto-increment from 1 for every emitted extraction.
- Attributes are converted from serde_json::Value::Object into
  AttributeMap, with scalar values becoming AttributeValue::Scalar
  and arrays becoming AttributeValue::List. Nested objects and
  nulls are dropped (lenient, matches Python).
- Sorting is stable so duplicate indices keep source order.

Tests mirror tests/resolver_test.py::ExtractOrderedEntitiesTest:
- valid_input_sorts_by_index
- empty_input_returns_empty
- mixed_index_order_sorts_stably
- missing_index_key_drops_the_extraction
- all_indices_missing_returns_empty_with_index_mode
- duplicate_indices_preserve_source_order
- negative_indices_sort_ascending
- index_without_data_key_ignored
- no_index_suffix_auto_increments
- attributes_are_attached_to_extractions
- non_integer_index_is_error
- non_scalar_extraction_text_is_error
- number_extraction_text_is_stringified
- end_to_end_parse_then_extract (full FormatHandler → records path
  on realistic fenced JSON output)

Verified: cargo test (88 total passed — 20 core + 19 tokenizer +
15 aligner + 34 format), cargo clippy --all-targets --all-features
-- -D warnings clean.

+++ rust/crates/langextract-core/src/data.rs
@@ -182,8 +182,14 @@ pub struct Extraction {
     pub alignment_status: Option<AlignmentStatus>,
 
     /// Index of this extraction in the output list.
+    ///
+    /// Signed because Python's `extraction_index` has no lower bound
+    /// and tests exercise negative indices for "lower priority" or
+    /// "placeholder" records. Resolvers that assign their own
+    /// positive, monotonically increasing indices can ignore the
+    /// sign.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub extraction_index: Option<usize>,
+    pub extraction_index: Option<i64>,
 
     /// Index of the group this extraction belongs to (used for co-referring
     /// extractions from the same output record).
